+++ src/add.js
@@ -1,33 +1,46 @@
+// Modules
+// =============================================================================
+var util = require('./util');
+
 // Exports
 // =============================================================================
-module.exports = function(elms, classNames) {
-    elms       = typeof elms === 'string' ? document.querySelectorAll(elms) : elms && elms.length ? elms : [elms];
-    classNames = classNames instanceof Array ? classNames.map(function(name){ return name.trim();}) : classNames.trim().replace(/\s+/g, ' ').split(' ');
+module.exports = function(target, classNames) {
+    elms = typeof target === 'string' ? document.querySelectorAll(target) : util.isIterableObj(target) ? target : [target];
 
-    function addClassNames(elm) {
-        // Native
-        if (elm.classList) {
-            classNames.forEach(function(className) {
-                elm.classList.add(className);
-            });
+    function addClassNames(elm, classNames) {
+        // Convert to array and trim values
+        classNames = util.classNamesToArray(classNames);
+
+        if (classNames && classNames.length) {
+            // Native
+            if (elm.classList) {
+                classNames.forEach(function(className) {
+                    elm.classList.add(className);
+                });
+            }
+            // Legacy
+            else {
+                var elmClasses = elm.className.split(' ');
+
+                classNames.forEach(function(className) {
+                    if (elmClasses.indexOf(className) === -1) {
+                        elmClasses.push(className);
+                    }
+                });
+
+                elm.className = elmClasses.join(' ');
+            }
         }
-        // Legacy
+    }
+
+    for (var i = 0, len = elms.length; i < len; i++) {
+        if (classNames instanceof Function) {
+            addClassNames(elms[i], classNames(elms[i], i));
+        }
         else {
-            var elmClasses = elm.className.split(' ');
-
-            classNames.forEach(function(className) {
-                if (elmClasses.indexOf(className) === -1) {
-                    elmClasses.push(className);
-                }
-            });
-
-            elm.className = elmClasses.join(' ');
+            addClassNames(elms[i], classNames);
         }
     }
 
-    for (var i = 0; i < elms.length; i++) {
-        addClassNames(elms[i]);
-    }
-
-    return elms;
+    return elms.length === 1 ? elms[0] : elms;
 };